+++ changelog.md
@@ -30,3 +30,12 @@
 
 - Removed sloppy code: parsing and validation of configuration is now done in one [place](https://github.com/tweedegolf/storage-abstraction/blob/master/src/util.ts).
 - Removed jasmine-ts dependency
+
+# 1.4.0
+
+- Replaced `introspect()` by `getConfiguration()`, `getOptions()` and `getType()`
+- Added adapter class for BackBlaze B2
+- Made configuration is more strict and generic
+- Removed option to create a new Storage without configuration (StorageLocal)
+- Added default storage options that can be overruled or extended by the options in the config object or url
+- Made slugify optional and turned it off by default on StorageLocal